Docker 환경 적용 + Review 수, 한글 지원 여부 확인

+++ backend/src/main/java/SteamGame/recommend/DTO/SteamDTO.java
@@ -39,7 +39,7 @@ public class SteamDTO {
     @Getter @Setter
     @JsonIgnoreProperties(ignoreUnknown = true)
     public static class SteamApp {
-        private int appid;
+        private Long appid;
         private String name;
         private String shortDescription;
         private String headerImage;
